+++ OpenHD/ohd_common/inc/openhd_led_pi.hpp
@@ -11,7 +11,7 @@
 
 #include "openhd_platform.hpp"
 #include "openhd_spdlog.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 
 // NOTE: Some PI's allow toggling both the red and green led
