+++ OpenHD/ohd_common/inc/openhd_settings_persistent.hpp
@@ -8,7 +8,7 @@
 #include <utility>
 
 #include "include_json.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 
 /**
