+++ OpenHD/ohd_interface/inc/wifi_card.hpp
@@ -8,7 +8,7 @@
 #include "openhd_platform.hpp"
 #include "openhd_settings_directories.hpp"
 #include "openhd_settings_persistent.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 #include "validate_settings_helper.h"
 #include "wifi_channel.h"
