+++ OpenHD/ohd_telemetry/src/internal/OnboardComputerStatus.hpp
@@ -11,7 +11,9 @@
 
 #include "mav_include.h"
 #include "openhd_spdlog.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
+#include "openhd_util_filesystem.hpp"
+#include <regex>
 
 // https://mavlink.io/en/messages/common.html#ONBOARD_COMPUTER_STATUS
 // used to be a custom message for a short amount of time.
