+++ OpenHD/ohd_video/inc/openhd-rpi-os-configure-vendor-cam.hpp
@@ -11,7 +11,7 @@
 
 #include "openhd_platform.hpp"
 #include "openhd_spdlog.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 
 // Helper to reconfigure the rpi os for the different camera types
