+++ OpenHD/ohd_video/inc/v_validate_settings.h
@@ -6,6 +6,7 @@
 #define OPENHD_OPENHD_OHD_VIDEO_INC_V_VALIDATE_SETTINGS_H_
 
 #include <optional>
+#include <regex>
 
 #include "openhd_spdlog.hpp"
