+++ OpenHD/ohd_video/src/air_recording_helper.hpp
@@ -8,7 +8,7 @@
 #include <sstream>
 #include <string>
 
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 
 namespace openhd::video{
