+++ OpenHD/ohd_video/src/camera_discovery.cpp
@@ -12,7 +12,7 @@
 #include "camera.hpp"
 #include "camera_discovery_helper.hpp"
 #include "libcamera_detect.hpp"
-#include "openhd_util.hpp"
+#include "openhd_util.h"
 #include "openhd_util_filesystem.hpp"
 
 // annoying linux platform specifics
